# Make mixgraph easier to use